Moving primitive types to interfaces
