+++ fuel-core-interfaces/Cargo.toml
@@ -15,16 +15,21 @@ anyhow = "1.0"
 async-trait = "0.1"
 fuel-asm = "0.3"
 fuel-storage = "0.1"
-fuel-tx = { version = "0.7", features = ["serde-types"] }
-fuel-types = { version = "0.3", features = ["serde-types"] }
-fuel-vm = "0.6"
+fuel-tx = { version = "0.7", default-features = false }
+fuel-types = { version = "0.3", default-features = false }
+fuel-vm = { version = "0.6", default-features = false }
+fuel-crypto = { version = "0.4", default-features = false}
 futures = "0.3"
 lazy_static = "1.4"
-parking_lot = "0.11"
+parking_lot = "0.12"
 thiserror = "1.0"
 tokio = { version = "1.14", features = ["full"] }
+derive_more = { version = "0.99" }
+serde = { version = "1.0", features = ["derive"], optional = true }
+chrono = "0.4"
 
 [dev-dependencies]
 
 [features]
 test_helpers = []
+serde-types = ["serde","fuel-tx/serde-types", "fuel-types/serde-types","fuel-vm/serde-types"]
