+++ fuel-core-interfaces/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod db;
 pub mod txpool;
+pub mod models;
