+++ fuel-core-interfaces/src/models/mod.rs
@@ -1,0 +1,7 @@
+mod block;
+mod block_height;
+mod coin;
+
+pub use block::{FuelBlock, FuelBlockDb, FuelBlockHeader};
+pub use block_height::BlockHeight;
+pub use coin::{Coin, CoinStatus};
