+++ fuel-core/Cargo.toml
@@ -19,7 +19,11 @@ test = false
 
 [dependencies]
 anyhow = "1.0"
-async-graphql = { version = "=2.9", features = ["chrono", "chrono-tz", "tracing"] }
+async-graphql = { version = "=2.9", features = [
+    "chrono",
+    "chrono-tz",
+    "tracing",
+] }
 async-trait = "0.1"
 axum = { version = "0.4" }
 bincode = "1.3"
@@ -29,7 +33,9 @@ derive_more = { version = "0.99" }
 dirs = "3.0"
 env_logger = "0.9"
 fuel-asm = { version = "0.3", features = ["serde-types"] }
-fuel-core-interfaces = { path = "../fuel-core-interfaces", version = "0.5.0" }
+fuel-core-interfaces = { path = "../fuel-core-interfaces", version = "0.5.0", features = [
+    "serde-types",
+] }
 fuel-crypto = { version = "0.3" }
 fuel-merkle = "0.1"
 fuel-storage = { version = "0.1" }
@@ -64,8 +70,16 @@ uuid = { version = "0.8", features = ["v4"] }
 
 [dev-dependencies]
 assert_matches = "1.5"
-fuel-tx = { version = "0.7", features = ["serde-types", "builder", "internals"] }
-fuel-vm = { version = "0.6", features = ["serde-types", "random", "test-helpers"] }
+fuel-tx = { version = "0.7", features = [
+    "serde-types",
+    "builder",
+    "internals",
+] }
+fuel-vm = { version = "0.6", features = [
+    "serde-types",
+    "random",
+    "test-helpers",
+] }
 insta = "1.8"
 
 [features]
