+++ fuel-core/src/database/metadata.rs
@@ -1,6 +1,6 @@
 use crate::database::columns::METADATA;
 use crate::database::Database;
-use crate::model::fuel_block::BlockHeight;
+use crate::model::BlockHeight;
 use crate::state::Error;
 
 const CHAIN_NAME_KEY: &[u8] = b"chain_name";
