+++ fuel-core/src/schema/coin.rs
@@ -1,6 +1,6 @@
 use crate::coin_query::{random_improve, SpendQueryElement};
 use crate::database::{Database, KvStoreError};
-use crate::model::coin::{Coin as CoinModel, CoinStatus};
+use crate::model::{Coin as CoinModel, LocalCoinStatus};
 use crate::schema::scalars::{Address, AssetId, UtxoId, U64};
 use crate::state::IterDirection;
 use async_graphql::InputObject;
@@ -8,6 +8,7 @@ use async_graphql::{
     connection::{query, Connection, Edge, EmptyFields},
     Context, Object,
 };
+use fuel_core_interfaces::models::CoinStatus;
 use fuel_storage::Storage;
 use fuel_tx::consts::MAX_INPUTS;
 use itertools::Itertools;
@@ -36,8 +37,8 @@ impl Coin {
         self.1.maturity.into()
     }
 
-    async fn status(&self) -> CoinStatus {
-        self.1.status
+    async fn status(&self) -> LocalCoinStatus {
+        self.1.status.into()
     }
 
     async fn block_created(&self) -> U64 {
